doc: use fixedbootstrap theme from probably
The bootstrap theme for sphinx has some issues that I fixed in the
"probably" library [1] with a custom theme that inherits from it.
When/if the offical bootstrap theme is fixed, the changes can be
reversed easily.

For now, these changes make the API documentation readable with added
spacing and highlights.

[1]: https://github.com/Philipp15b/probably